added tron2 which has the following difference with tron
+the game allows a player to take over an AI and thus join midgame

+the game now has a maximum game speed thus allowing the players too keep
playing this still has to be tested on the hexxx itself and adjusted

+the game speed has been slightly randomised to make it harder

+the game implemented a 50% chance to jump if both buttons are pressed at the same time.
this allowes you to jump over another player. or go faster

+the game now has an interpolated step which might look like the game runs more fluid or faster. still has to
be checked on the hexxx

later editions of the game will have to be redisigned so it doesn't work on the raw game buffers anymore.
this will allow for features like dropping a bomb/ peice of the tail aswell as snake like growing.
Because we then know the location and lenght of the full tron trail.